requirements.txt
Package                   Version
------------------------- ---------
altgraph                  0.17.3
astroid                   2.15.6
auto-py-to-exe            2.37.0
av                        10.0.0
bottle                    0.12.25
bottle-websocket          0.2.9
certifi                   2023.7.22
cffi                      1.15.1
charset-normalizer        3.2.0
click                     8.1.7
colorama                  0.4.6
comtypes                  1.2.0
cups                      0.0.6
decorator                 4.4.2
dill                      0.3.7
Eel                       0.16.0
future                    0.18.3
gevent                    23.7.0
gevent-websocket          0.10.1
greenlet                  2.0.2
gTTS                      2.3.2
idna                      3.4
imageio                   2.31.2
imageio-ffmpeg            0.4.8
isort                     5.12.0
keyboard                  0.13.5
lazy-object-proxy         1.9.0
mccabe                    0.7.0
MouseInfo                 0.1.3
moviepy                   1.0.3
numpy                     1.25.2
pefile                    2023.2.7
Pillow                    10.0.0
pip                       23.3.1
platformdirs              3.10.0
proglog                   0.1.10
PyAudio                   0.2.13
PyAutoGUI                 0.9.54
pycparser                 2.21
pydub                     0.25.1
pygame                    2.5.1
PyGetWindow               0.0.9
pyinstaller               5.13.0
pyinstaller-hooks-contrib 2023.7
pylint                    2.17.5
PyMsgBox                  1.0.9
pyparsing                 3.1.1
pyperclip                 1.8.2
pypiwin32                 223
PyRect                    0.2.0
PyScreeze                 0.1.30
pyttsx3                   2.90
pytweening                1.0.7
pywin32                   306
pywin32-ctypes            0.2.2
requests                  2.31.0
SpeechRecognition         3.10.1
tk                        0.1.0
tkvideoplayer             2.3
tomlkit                   0.12.1
tqdm                      4.66.1
typing_extensions         4.8.0
urllib3                   2.0.4
Wave                      0.0.2
whichcraft                0.6.1
wrapt                     1.15.0
zope.event                5.0
zope.interface            6.0
